[ISSUE#3179]Replace the NotifyCenter with new refactor in the core and test module.
