+++ core/src/main/java/com/alibaba/nacos/core/cluster/MemberChangeListener.java
@@ -16,15 +16,16 @@
 
 package com.alibaba.nacos.core.cluster;
 
-import com.alibaba.nacos.core.notify.Event;
-import com.alibaba.nacos.core.notify.listener.Subscribe;
+import com.alibaba.nacos.common.notify.Event;
+import com.alibaba.nacos.common.notify.listener.Subscriber;
 
 /**
  * Node change listeners.
  *
  * @author <a href="mailto:liaochuntao@live.com">liaochuntao</a>
  */
-public interface MemberChangeListener extends Subscribe<MembersChangeEvent> {
+@SuppressWarnings("PMD.AbstractClassShouldStartWithAbstractNamingRule")
+public abstract class MemberChangeListener extends Subscriber<MembersChangeEvent> {
     
     /**
      * return NodeChangeEvent.class info.
@@ -32,7 +33,7 @@ public interface MemberChangeListener extends Subscribe<MembersChangeEvent> {
      * @return {@link MembersChangeEvent#getClass()}
      */
     @Override
-    default Class<? extends Event> subscribeType() {
+    public Class<? extends Event> subscribeType() {
         return MembersChangeEvent.class;
     }
     
@@ -42,7 +43,7 @@ public interface MemberChangeListener extends Subscribe<MembersChangeEvent> {
      * @return default value is {@link Boolean#TRUE}
      */
     @Override
-    default boolean ignoreExpireEvent() {
+    public boolean ignoreExpireEvent() {
         return true;
     }
 }
