+++ core/src/main/java/com/alibaba/nacos/core/cluster/ServerMemberManager.java
@@ -29,9 +29,9 @@ import com.alibaba.nacos.common.utils.ConcurrentHashSet;
 import com.alibaba.nacos.common.utils.ExceptionUtil;
 import com.alibaba.nacos.common.utils.VersionUtils;
 import com.alibaba.nacos.core.cluster.lookup.LookupFactory;
-import com.alibaba.nacos.core.notify.Event;
-import com.alibaba.nacos.core.notify.NotifyCenter;
-import com.alibaba.nacos.core.notify.listener.Subscribe;
+import com.alibaba.nacos.common.notify.Event;
+import com.alibaba.nacos.common.notify.NotifyCenter;
+import com.alibaba.nacos.common.notify.listener.Subscriber;
 import com.alibaba.nacos.core.utils.ApplicationUtils;
 import com.alibaba.nacos.core.utils.Commons;
 import com.alibaba.nacos.core.utils.Constants;
@@ -166,7 +166,7 @@ public class ServerMemberManager implements ApplicationListener<WebServerInitial
         
         // The address information of this node needs to be dynamically modified
         // when registering the IP change of this node
-        NotifyCenter.registerSubscribe(new Subscribe<InetUtils.IPChangeEvent>() {
+        NotifyCenter.registerSubscriber(new Subscriber<InetUtils.IPChangeEvent>() {
             @Override
             public void onEvent(InetUtils.IPChangeEvent event) {
                 String newAddress = event.getNewIp() + ":" + port;
