+++ core/src/main/java/com/alibaba/nacos/core/code/StartingSpringApplicationRunListener.java
@@ -21,7 +21,7 @@ import com.alibaba.nacos.common.executor.NameThreadFactory;
 import com.alibaba.nacos.common.executor.ThreadPoolManager;
 import com.alibaba.nacos.common.http.HttpClientManager;
 import com.alibaba.nacos.core.file.WatchFileCenter;
-import com.alibaba.nacos.core.notify.NotifyCenter;
+import com.alibaba.nacos.common.notify.NotifyCenter;
 import com.alibaba.nacos.core.utils.ApplicationUtils;
 import com.alibaba.nacos.core.utils.DiskUtils;
 import com.alibaba.nacos.core.utils.InetUtils;
