+++ core/src/main/java/com/alibaba/nacos/core/distributed/raft/NacosStateMachine.java
@@ -30,7 +30,7 @@ import com.alibaba.nacos.consistency.snapshot.Reader;
 import com.alibaba.nacos.consistency.snapshot.SnapshotOperation;
 import com.alibaba.nacos.consistency.snapshot.Writer;
 import com.alibaba.nacos.core.distributed.raft.utils.JRaftUtils;
-import com.alibaba.nacos.core.notify.NotifyCenter;
+import com.alibaba.nacos.common.notify.NotifyCenter;
 import com.alibaba.nacos.core.utils.Loggers;
 import com.alipay.sofa.jraft.Closure;
 import com.alipay.sofa.jraft.Iterator;
