+++ core/src/main/java/com/alibaba/nacos/core/distributed/raft/RaftEvent.java
@@ -16,7 +16,7 @@
 
 package com.alibaba.nacos.core.distributed.raft;
 
-import com.alibaba.nacos.core.notify.SlowEvent;
+import com.alibaba.nacos.common.notify.SlowEvent;
 
 import java.util.Collections;
 import java.util.List;
@@ -27,7 +27,7 @@ import java.util.List;
  * @author <a href="mailto:liaochuntao@live.com">liaochuntao</a>
  */
 @SuppressWarnings("all")
-public class RaftEvent implements SlowEvent {
+public class RaftEvent extends SlowEvent {
     
     private static final long serialVersionUID = -4304258594602886451L;
