+++ core/src/main/java/com/alibaba/nacos/core/utils/InetUtils.java
@@ -16,8 +16,8 @@
 
 package com.alibaba.nacos.core.utils;
 
-import com.alibaba.nacos.core.notify.NotifyCenter;
-import com.alibaba.nacos.core.notify.SlowEvent;
+import com.alibaba.nacos.common.notify.NotifyCenter;
+import com.alibaba.nacos.common.notify.SlowEvent;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -229,7 +229,7 @@ public class InetUtils {
      * {@link com.alibaba.nacos.core.cluster.ServerMemberManager} is listener.
      */
     @SuppressWarnings({"PMD.ClassNamingShouldBeCamelRule", "checkstyle:AbbreviationAsWordInName"})
-    public static class IPChangeEvent implements SlowEvent {
+    public static class IPChangeEvent extends SlowEvent {
         
         private String oldIp;
