+++ naming/src/main/java/com/alibaba/nacos/naming/consistency/persistent/raft/RaftPeerSet.java
@@ -21,7 +21,7 @@ import com.alibaba.nacos.core.cluster.Member;
 import com.alibaba.nacos.core.cluster.MemberChangeListener;
 import com.alibaba.nacos.core.cluster.MembersChangeEvent;
 import com.alibaba.nacos.core.cluster.ServerMemberManager;
-import com.alibaba.nacos.core.notify.NotifyCenter;
+import com.alibaba.nacos.common.notify.NotifyCenter;
 import com.alibaba.nacos.core.utils.ApplicationUtils;
 import com.alibaba.nacos.naming.misc.HttpClient;
 import com.alibaba.nacos.naming.misc.Loggers;
@@ -53,7 +53,7 @@ import java.util.concurrent.atomic.AtomicLong;
  */
 @Component
 @DependsOn("ProtocolManager")
-public class RaftPeerSet implements MemberChangeListener {
+public class RaftPeerSet extends MemberChangeListener {
     
     private final ServerMemberManager memberManager;
     
@@ -75,7 +75,7 @@ public class RaftPeerSet implements MemberChangeListener {
     
     @PostConstruct
     public void init() {
-        NotifyCenter.registerSubscribe(this);
+        NotifyCenter.registerSubscriber(this);
         changePeers(memberManager.allMembers());
     }
