+++ test/src/test/java/com/alibaba/nacos/test/core/cluster/ServerMemberManager_ITCase.java
@@ -21,9 +21,9 @@ import com.alibaba.nacos.core.cluster.MembersChangeEvent;
 import com.alibaba.nacos.core.cluster.MemberUtils;
 import com.alibaba.nacos.core.cluster.NodeState;
 import com.alibaba.nacos.core.cluster.ServerMemberManager;
-import com.alibaba.nacos.core.notify.Event;
-import com.alibaba.nacos.core.notify.NotifyCenter;
-import com.alibaba.nacos.core.notify.listener.Subscribe;
+import com.alibaba.nacos.common.notify.Event;
+import com.alibaba.nacos.common.notify.NotifyCenter;
+import com.alibaba.nacos.common.notify.listener.Subscriber;
 import com.alibaba.nacos.core.utils.ApplicationUtils;
 import com.alibaba.nacos.core.utils.Constants;
 import org.junit.After;
@@ -108,7 +108,7 @@ public class ServerMemberManager_ITCase {
         AtomicInteger integer = new AtomicInteger(0);
         CountDownLatch latch = new CountDownLatch(1);
         
-        NotifyCenter.registerSubscribe(new Subscribe<MembersChangeEvent>() {
+        NotifyCenter.registerSubscriber(new Subscriber<MembersChangeEvent>() {
             @Override
             public void onEvent(MembersChangeEvent event) {
                 integer.incrementAndGet();
@@ -141,7 +141,7 @@ public class ServerMemberManager_ITCase {
         AtomicReference<Collection<Member>> healthMembers = new AtomicReference<>();
         CountDownLatch first = new CountDownLatch(1);
         CountDownLatch second = new CountDownLatch(1);
-        NotifyCenter.registerSubscribe(new Subscribe<MembersChangeEvent>() {
+        NotifyCenter.registerSubscriber(new Subscriber<MembersChangeEvent>() {
             @Override
             public void onEvent(MembersChangeEvent event) {
                 System.out.println(event);
